+++ src/alloc.h
@@ -1,7 +1,9 @@
 #ifndef MEMPOOL_H
 #define MEMPOOL_H
 
-void *mp_alloc(unsigned size);
+#include <sys/types.h>
+
+void *mp_alloc(size_t size);
 void mp_free(void *buffer);
 void mp_force_free();
